rename to `@theia/request`
