+++ dev-packages/cli/package.json
@@ -35,7 +35,7 @@
     "@theia/ffmpeg": "1.25.0",
     "@theia/localization-manager": "1.25.0",
     "@theia/ovsx-client": "1.25.0",
-    "@theia/request-service": "1.25.0",
+    "@theia/request": "1.25.0",
     "@types/chai": "^4.2.7",
     "@types/mocha": "^5.2.7",
     "@types/node-fetch": "^2.5.7",
