+++ packages/core/README.md
@@ -89,9 +89,9 @@ export class SomeClass {
     - `@theia/application-package` (from [`@theia/application-package@1.25.0`](https://www.npmjs.com/package/@theia/application-package/v/1.25.0))
     - `@theia/application-package/lib/api` (from [`@theia/application-package@1.25.0`](https://www.npmjs.com/package/@theia/application-package/v/1.25.0))
     - `@theia/application-package/lib/environment` (from [`@theia/application-package@1.25.0`](https://www.npmjs.com/package/@theia/application-package/v/1.25.0))
-    - `@theia/request-service` (from [`@theia/request-service@1.25.0`](https://www.npmjs.com/package/@theia/request-service/v/1.25.0))
-    - `@theia/request-service/lib/proxy` (from [`@theia/request-service@1.25.0`](https://www.npmjs.com/package/@theia/request-service/v/1.25.0))
-    - `@theia/request-service/lib/node-request-service` (from [`@theia/request-service@1.25.0`](https://www.npmjs.com/package/@theia/request-service/v/1.25.0))
+    - `@theia/request` (from [`@theia/request@1.25.0`](https://www.npmjs.com/package/@theia/request/v/1.25.0))
+    - `@theia/request/lib/proxy` (from [`@theia/request@1.25.0`](https://www.npmjs.com/package/@theia/request/v/1.25.0))
+    - `@theia/request/lib/node-request-service` (from [`@theia/request@1.25.0`](https://www.npmjs.com/package/@theia/request/v/1.25.0))
     - `fs-extra` (from [`fs-extra@^4.0.2`](https://www.npmjs.com/package/fs-extra))
     - `fuzzy` (from [`fuzzy@^0.1.3`](https://www.npmjs.com/package/fuzzy))
     - `inversify` (from [`inversify@^5.1.1`](https://www.npmjs.com/package/inversify))
