+++ packages/core/package.json
@@ -17,7 +17,7 @@
     "@phosphor/virtualdom": "1",
     "@phosphor/widgets": "1",
     "@theia/application-package": "1.25.0",
-    "@theia/request-service": "1.25.0",
+    "@theia/request": "1.25.0",
     "@types/body-parser": "^1.16.4",
     "@types/cookie": "^0.3.3",
     "@types/dompurify": "^2.2.2",
@@ -104,9 +104,9 @@
         "@theia/application-package",
         "@theia/application-package/lib/api",
         "@theia/application-package/lib/environment",
-        "@theia/request-service",
-        "@theia/request-service/lib/proxy",
-        "@theia/request-service/lib/node-request-service",
+        "@theia/request",
+        "@theia/request/lib/proxy",
+        "@theia/request/lib/node-request-service",
         "fs-extra",
         "fuzzy",
         "inversify",
