--- packages/core/shared/@theia/request-service/index.d.ts
@@ -1,1 +1,0 @@
-export * from '@theia/request-service';
