--- packages/core/shared/@theia/request-service/index.js
@@ -1,1 +1,0 @@
-module.exports = require('@theia/request-service');
