+++ packages/core/shared/@theia/request/index.d.ts
@@ -1,0 +1,1 @@
+export * from '@theia/request';
