+++ packages/core/src/electron-node/request/electron-backend-request-module.ts
@@ -15,7 +15,7 @@
  ********************************************************************************/
 
 import { ContainerModule } from 'inversify';
-import { RequestService } from '@theia/request-service';
+import { RequestService } from '@theia/request';
 import { ElectronBackendRequestService } from './electron-backend-request-service';
 
 export default new ContainerModule(bind => {
