+++ packages/vsx-registry/src/node/vsx-registry-backend-module.ts
@@ -21,7 +21,7 @@ import { OVSXClientProvider, createOVSXClient } from '../common/ovsx-client-prov
 import { VSXEnvironment, VSX_ENVIRONMENT_PATH } from '../common/vsx-environment';
 import { VSXEnvironmentImpl } from './vsx-environment-impl';
 import { ConnectionHandler, JsonRpcConnectionHandler } from '@theia/core';
-import { RequestService } from '@theia/core/shared/@theia/request-service';
+import { RequestService } from '@theia/core/shared/@theia/request';
 
 export default new ContainerModule(bind => {
     bind<OVSXClientProvider>(OVSXClientProvider).toDynamicValue(ctx => {
